requirements.txt
# Passive Income - Bitcoin Trading with TDA
# Python 3.10 required
# Install PyTorch separately: conda install pytorch torchvision torchaudio pytorch-cuda=12.1 -c pytorch -c nvidia

# TDA Libraries
giotto-tda==0.6.2
giotto-ph==0.2.4
pyflagser==0.4.7

# Data Analysis
pandas==2.3.3
numpy==1.26.4
scipy==1.15.3
scikit-learn==1.3.2

# Visualization
matplotlib==3.10.8
seaborn==0.13.2
plotly==6.5.0
pillow==12.0.0

# Utilities
pyyaml>=6.0
tqdm==4.67.1
requests==2.32.5
python-dateutil==2.9.0.post0

# Development
ipython==8.37.0
ipywidgets==8.1.8

# Graph/Network
igraph==1.0.0
